feat: add showcase example to main Avatar docs
