+++ src/components/Avatar/Examples.stories.tsx
@@ -4,7 +4,9 @@ import { Sparkles as SparklesDrei } from '@react-three/drei';
 import type { Meta } from '@storybook/react';
 import { Avatar as AvatarWrapper, CAMERA } from 'src/components/Avatar';
 import { getStoryAssetPath } from 'src/services';
+import { ignoreArgTypesOnExamples, emotions } from 'src/services/Stories.service';
 import { AvatarProps } from './Avatar.component';
+import { Static } from './Avatar.stories';
 
 const Avatar = (args: AvatarProps) => <AvatarWrapper {...args} />;
 
@@ -16,38 +18,6 @@ const meta: Meta<typeof Avatar> = {
 };
 
 export default meta;
-
-const disableTable = { table: { disable: true } };
-const keysToIgnore = [
-  'modelSrc',
-  'animationSrc',
-  'poseSrc',
-  'halfBody',
-  'shadows',
-  'cameraTarget',
-  'cameraInitialDistance',
-  'style',
-  'idleRotation',
-  'emotion',
-  'background',
-  'capture',
-  'loader',
-  'dpr',
-  'className',
-  'headMovement',
-  'cameraZoomTarget',
-  'bloom',
-  'onLoadedEffect',
-  'onLoadedAnimation',
-  'children'
-];
-const ignoreArgTypesOnExamples = keysToIgnore.reduce(
-  (acc, key) => {
-    acc[key] = disableTable;
-    return acc;
-  },
-  {} as Record<string, typeof disableTable>
-);
 
 export const FloatingSparkles: StoryFn<typeof SparklesDrei> = (args) => (
   <Avatar
@@ -104,68 +74,19 @@ SpawnEffectAndAnimation.argTypes = {
   animationSrc: { control: { disable: false } }
 };
 
-/**
- * https://models.readyplayer.me/64d61e9e17883fd73ebe5eb7.glb
- * https://models.readyplayer.me/64d61f67f0367d07504924be.glb
- * https://models.readyplayer.me/64d62255f0367d0750492913.glb
- * https://models.readyplayer.me/64d622ce17883fd73ebe64cd.glb
- * https://models.readyplayer.me/64d6235e2d3bea6e4267b01d.glb
- * https://models.readyplayer.me/64d5d4eb651a0d350005c672.glb
- */
-
-export const modelPresets = {
-  one: 'https://models.readyplayer.me/64d61e9e17883fd73ebe5eb7.glb?morphTargets=Default,ARKit&lod=0',
-  two: 'https://models.readyplayer.me/64d61f67f0367d07504924be.glb',
-  three: 'https://models.readyplayer.me/64d62255f0367d0750492913.glb',
-  four: 'https://models.readyplayer.me/64d622ce17883fd73ebe64cd.glb',
-  five: 'https://models.readyplayer.me/64d6235e2d3bea6e4267b01d.glb',
-  six: 'https://models.readyplayer.me/64d5d4eb651a0d350005c672.glb'
-};
-export const Showcase: StoryFn<typeof Avatar> = (args) => (
-  <Avatar {...args}>
-    <Sparkles color="white" count={50} opacity={0.9} scale={5} size={0.5} speed={0.35} />
-  </Avatar>
-);
-Showcase.args = {
-  onLoadedAnimation: {
-    src: getStoryAssetPath('M_Standing_Idle_Variations_002.fbx'),
-    loop: 1
-  },
-  cameraTarget: CAMERA.TARGET.FULL_BODY.FEMALE,
+export const Posing: StoryFn<typeof Avatar> = (args) => <Avatar {...args} />;
+Posing.args = {
+  ...Static.args,
+  modelSrc: getStoryAssetPath('male.glb'),
+  poseSrc: getStoryAssetPath('male-pose-standing.glb'),
+  cameraTarget: CAMERA.TARGET.FULL_BODY.MALE,
   cameraInitialDistance: CAMERA.CONTROLS.FULL_BODY.MAX_DISTANCE,
-  modelSrc: 'https://models.readyplayer.me/64d61e9e17883fd73ebe5eb7.glb?morphTargets=Default,ARKit&lod=0',
-  animationSrc: getStoryAssetPath('M_Standing_Idle_001.fbx'),
-  bloom: {
-    luminanceThreshold: 1.0,
-    luminanceSmoothing: 1.0,
-    mipmapBlur: true,
-    kernelSize: 1,
-    intensity: 1,
-    materialIntensity: 6
-  },
-  style: { background: '#282038' },
-  dpr: 2,
-  ambientLightColor: '#ffffff',
-  dirLightColor: '#ffffff',
-  spotLightColor: '#adbfe5',
-  ambientLightIntensity: 0,
-  dirLightIntensity: 2.2,
-  spotLightIntensity: 0.5,
-  environment: 'apartment',
-  shadows: true,
-  emotion: {
-    jawOpen: 0.1,
-    mouthSmileLeft: 0.2,
-    mouthSmileRight: 0.1,
-    mouthPressLeft: 0.1,
-    cheekSquintLeft: 0.3,
-    eyeLookOutLeft: 0.6,
-    eyeLookInRight: 0.6,
-    mouthDimpleLeft: 0.3
-  }
+  /* eslint-disable no-console */
+  onLoaded: () => console.info('EVENT: posing avatar loaded'),
+  onLoading: () => console.info('EVENT: loading posing avatar'),
+  /* eslint-enable no-console */
+  emotion: emotions.smile
 };
-Showcase.argTypes = {
-  onLoadedAnimation: { control: { disable: false } },
-  animationSrc: { control: { disable: false } },
-  modelSrc: { options: Object.values(modelPresets), control: { type: 'select' } }
+Posing.argTypes = {
+  headMovement: { control: false }
 };
